+++ src/pages/messages/messages.template.pug
@@ -3,7 +3,7 @@ div(class='container')
         div(class='list-header')
             template(data-component="searchInput")
             div(class='profile-container')
-                a(href='' class='profile') Профиль >
+                a(href='' class='profile' id='profile') Профиль >
 
         div(class='single-chat-container')
             div(class='avatar-container')
